Add support for suite-level callback on timeout
This allows the suite to be able to collect whatever extra information
it might need to collect before panicing by adding an `OnTimeout(*C)`
function to the test suite.

Example would be Docker's test suites which keeps multiple docker
daemons running in the background as `exec.Command`. On occasions where
a timeout occurs, it is helpful to be able to trigger some actions such
as triggering a stack dump on the running daemons.

You cannot use `OnTimeout` to recover from a timeout.
`OnTimeout` is also called with a timeout, currently set to the check
suite's timeout value to prevent calls in `OnTimeout` from freezing the
system.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>